refactor: 게임 진행 중 상태 공통 로직 구현
